ShoeScan/requirements.txt
streamlit==1.54.0
google-genai==1.62.0
python-dotenv==1.2.1
